requirements.txt
google-adk>=0.3.0
fastmcp>=0.4.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
google-api-python-client>=2.120.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
pydantic>=2.0.0
httpx>=0.27.0
